+++ PokemonBattleEngine/Packets/_PkmnHPChangedPacket.cs
@@ -10,8 +10,8 @@ namespace Kermalis.PokemonBattleEngine.Packets
     {
         PBETrainer PokemonTrainer { get; }
         PBEFieldPosition Pokemon { get; }
-        double OldHPPercentage { get; }
-        double NewHPPercentage { get; }
+        float OldHPPercentage { get; }
+        float NewHPPercentage { get; }
     }
     public sealed class PBEPkmnHPChangedPacket : IPBEPkmnHPChangedPacket
     {
@@ -22,10 +22,10 @@ namespace Kermalis.PokemonBattleEngine.Packets
         public PBEFieldPosition Pokemon { get; }
         public ushort OldHP { get; }
         public ushort NewHP { get; }
-        public double OldHPPercentage { get; }
-        public double NewHPPercentage { get; }
+        public float OldHPPercentage { get; }
+        public float NewHPPercentage { get; }
 
-        internal PBEPkmnHPChangedPacket(PBEBattlePokemon pokemon, ushort oldHP, double oldHPPercentage)
+        internal PBEPkmnHPChangedPacket(PBEBattlePokemon pokemon, ushort oldHP, float oldHPPercentage)
         {
             using (var ms = new MemoryStream())
             using (var w = new EndianBinaryWriter(ms, encoding: EncodingType.UTF16))
@@ -47,8 +47,8 @@ namespace Kermalis.PokemonBattleEngine.Packets
             Pokemon = r.ReadEnum<PBEFieldPosition>();
             OldHP = r.ReadUInt16();
             NewHP = r.ReadUInt16();
-            OldHPPercentage = r.ReadDouble();
-            NewHPPercentage = r.ReadDouble();
+            OldHPPercentage = r.ReadSingle();
+            NewHPPercentage = r.ReadSingle();
         }
     }
     public sealed class PBEPkmnHPChangedPacket_Hidden : IPBEPkmnHPChangedPacket
@@ -58,8 +58,8 @@ namespace Kermalis.PokemonBattleEngine.Packets
 
         public PBETrainer PokemonTrainer { get; }
         public PBEFieldPosition Pokemon { get; }
-        public double OldHPPercentage { get; }
-        public double NewHPPercentage { get; }
+        public float OldHPPercentage { get; }
+        public float NewHPPercentage { get; }
 
         public PBEPkmnHPChangedPacket_Hidden(PBEPkmnHPChangedPacket other)
         {
@@ -83,8 +83,8 @@ namespace Kermalis.PokemonBattleEngine.Packets
             Data = new ReadOnlyCollection<byte>(data);
             PokemonTrainer = battle.Trainers[r.ReadByte()];
             Pokemon = r.ReadEnum<PBEFieldPosition>();
-            OldHPPercentage = r.ReadDouble();
-            NewHPPercentage = r.ReadDouble();
+            OldHPPercentage = r.ReadSingle();
+            NewHPPercentage = r.ReadSingle();
         }
     }
 }
